Function to create square or heaxagonal planning units for your area of interest.

The code takes the bbox so the limits are the most important. The output inherits the crs from this sf object so ensure it is in the correct projection for your needs The code assumes that any Planning Unit whose centroids is over land will be removed. This approximates > 50% of the PU is landward.

Usage

SpatPlan_Get_PlanningUnits(
  Bndry,
  InnerB,
  CellArea = 1000,
  Shape = "hexagon",
  inverse = FALSE
)

Arguments

Bndry

An sf polygon object which outlines the limits of the study area.

InnerB

An sf multipolygon object which contains all the areas (ie land) that you wish to remove from the grid.

CellArea

The area in km you wish your resultant Planning Units to be.

Shape

Hexagon or Square

inverse

To invert the InnerB and keep the areas within the InnerB

Value

An sf object of planning units

Details

Some example options for hexagonal PU_size PU_size <- 5240 # km2 (~1 deg at equator) PU_size <- 2620 # km2 (~0.5 deg at equator) PU_size <- 669.9 # km2 (~0.25 deg at equator)

Written by Jason Everett (UQ/UNSW/CSIRO) Written: 15 December 2020 Updated: 8th April 2022