Function to recover data from Global Fishing Watch.

The code takes several parameters described below and return an sf object with gfw data aggregated or not (param compress)

Usage

splnr_get_gfw(
  region,
  start_date,
  end_date,
  temp_res,
  spat_res = "low",
  key = gfwr::gfw_auth(),
  cCRS = "EPSG:4326",
  compress = FALSE
)

Arguments

region

Region studied (character).

start_date

Start date (waited format : "%Y-%m-%d").

end_date

End date (waited format : "%Y-%m-%d").

temp_res

Temporal resolution ("daily","monthly","yearly").

spat_res

Spatial resolution ("low" for 0.1 degree, "high" for 0.01 degree).

key

Token for GFW API (see details GlobalFishingWatch_Examples vignette).

cCRS

The csr to which the sf will be returned (default = "EPSG:4326").

compress

Binary operator to compress (aggregate) the data per coordinates (default = FALSE).

Value

An sf object with gfw data.

Details

Written by Kilian Barreiro Written: February 2024

Examples

if (FALSE) {
Region <- "Coral Sea" # "Australia"
Type <- "Oceans" # "EEZ"

Shape <- "Hexagon" # "Shape of PUs
PU_size <- 10000 # km2

cCRS <- "ESRI:54009"

Bndry <- spatialplanr::splnr_get_boundary(Region, Type, cCRS)

landmass <- rnaturalearth::ne_countries(
 scale = "medium",
 returnclass = "sf"
) %>%
sf::st_transform(cCRS)

PUs <- spatialplanr::splnr_get_planningUnits(Bndry, landmass, PU_size, Shape)

gfw_data <- splnr_get_gfw('Australia', "2022-01-01", "2022-12-31", "yearly",
cCRS = cCRS, compress = TRUE)
}